pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "telethon==1.42.0",
]
